code restruct
